backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Weather API
requests==2.31.0
python-dotenv==1.0.0

# Google Sheets
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Validation
pydantic==2.5.0
email-validator==2.1.0
